lisa stuff

+++ s452/unpack_bplast_twinpeaks.cc
@@ -96,7 +96,7 @@ void EXT_PLASTIC_TP::Process_TAMEX(__data_src_t &__buffer, plastic_twinpeaks_ite
     uint32 tamex_header = 0;
     __buffer.peek_uint32(&tamex_header);
 
-    bool ongoing = ((tamex_header & 0xFF) == item.tamex_identifier) && (((tamex_header >> 24) & 0xFF) == 0) && ((((tamex_header >> 12) & 0xF) == 1) || (((tamex_header >> 12) & 0xF) == 0));
+    bool ongoing = ((tamex_header & 0xFF) == (uint32) item.tamex_identifier) && (((tamex_header >> 24) & 0xFF) == 0) && ((((tamex_header >> 12) & 0xF) == 1) || (((tamex_header >> 12) & 0xF) == 0));
     
     if (!ongoing)
     {
@@ -134,7 +134,7 @@ void EXT_PLASTIC_TP::Process_TAMEX(__data_src_t &__buffer, plastic_twinpeaks_ite
 
     uint32 tamex_begin = 0;
     __buffer.peek_uint32(&tamex_begin);
-    if (((tamex_begin >> 24) & 0xFF) != item.aa)
+    if (((tamex_begin >> 24) & 0xFF) != (uint32) item.aa)
     {
         std::cerr << "error in TAMEX format! 0xAA... word not found after fired amount!" << std::endl;
         std::cerr << "TAMEX word: " << std::hex << tamex_begin << std::endl;
@@ -245,12 +245,12 @@ void EXT_PLASTIC_TP::get_edges(__data_src_t &__buffer, plastic_twinpeaks_item &i
         uint32 place_holder = 0;
         __buffer.peek_uint32(&place_holder);
 
-        if (((place_holder >> 28) & 0xF) == item.six_f) // six_f is just 6..idk
+        if (((place_holder >> 28) & 0xF) == (uint32) item.six_f) // six_f is just 6..idk
         {
             __buffer.advance(4);
             continue;
         }
-        if (((place_holder >> 28) & 0xF) != item.six_f)
+        if (((place_holder >> 28) & 0xF) != (uint32) item.six_f)
         {
             item.written = false;
         }
